% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra_call_stack.R
\name{infra_call_stack}
\alias{infra_call_stack}
\title{Build call stack of function names for messages}
\usage{
infra_call_stack(child = NULL, parent = NULL, ...)
}
\arguments{
\item{child}{(optional, character) Name of the calling function. Default: NULL}

\item{parent}{(optional, character) Name of the parent function. Default: NULL}

\item{...}{(optional) Passes \code{parent} to functions that lack a \code{parent} parameter.}
}
\value{
A character string with the formatted call stack, or \code{NULL} if both \code{parent} and \code{child} are \code{NULL}.
}
\description{
Concatenates parent and child function names to improve message, warning, and error tracing.
}
\examples{
x <- infra_call_stack(
  parent = "parent_name",
  child = "child_function"
)

message(x)
}
\seealso{
Other infra:
\code{\link[=infra_get_wkt_column]{infra_get_wkt_column()}},
\code{\link[=infra_get_x_column]{infra_get_x_column()}},
\code{\link[=infra_get_xy_columns]{infra_get_xy_columns()}},
\code{\link[=infra_get_y_column]{infra_get_y_column()}},
\code{\link[=infra_guess_crs]{infra_guess_crs()}},
\code{\link[=infra_validate_df]{infra_validate_df()}},
\code{\link{infra_x_column_names}},
\code{\link{infra_y_column_names}}
}
\concept{infra}
