% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_x_column_names}
\alias{infra_x_column_names}
\title{Recognized longitude column name variants}
\format{
vector
}
\usage{
data(infra_x_column_names)
}
\description{
A character vector of recognized lowercase column name variants for longitude
coordinates, covering multiple languages.
}
\seealso{
Other infra:
\code{\link[=infra_call_stack]{infra_call_stack()}},
\code{\link[=infra_get_wkt_column]{infra_get_wkt_column()}},
\code{\link[=infra_get_x_column]{infra_get_x_column()}},
\code{\link[=infra_get_xy_columns]{infra_get_xy_columns()}},
\code{\link[=infra_get_y_column]{infra_get_y_column()}},
\code{\link[=infra_guess_crs]{infra_guess_crs()}},
\code{\link[=infra_validate_df]{infra_validate_df()}},
\code{\link{infra_y_column_names}}
}
\concept{infra}
\keyword{datasets}
